--- packages/react-examples/src/react-next/SpinButton/SpinButton.BasicWithEndPosition.Example.tsx
@@ -1,19 +1,0 @@
-import * as React from 'react';
-import { SpinButton } from '@fluentui/react-next/lib/SpinButton';
-import { Position } from '@fluentui/react-next/lib/Positioning';
-
-export const SpinButtonBasicWithEndPositionExample: React.FC = () => (
-  <div style={{ width: '400px' }}>
-    <SpinButton
-      defaultValue="0"
-      iconProps={{ iconName: 'Light' }}
-      label={'Basic SpinButton'}
-      labelPosition={Position.end}
-      min={0}
-      max={100}
-      step={1}
-      incrementButtonAriaLabel={'Increase value by 1'}
-      decrementButtonAriaLabel={'Decrease value by 1'}
-    />
-  </div>
-);
