--- packages/react-examples/src/react-next/SpinButton/SpinButton.BasicWithIconDisabled.Example.tsx
@@ -1,18 +1,0 @@
-import * as React from 'react';
-import { SpinButton } from '@fluentui/react-next/lib/SpinButton';
-
-export const SpinButtonBasicWithIconDisabledExample: React.FC = () => (
-  <div style={{ width: '400px' }}>
-    <SpinButton
-      defaultValue="0"
-      disabled={true}
-      iconProps={{ iconName: 'IncreaseIndentLegacy' }}
-      label={'Basic SpinButton:'}
-      min={0}
-      max={100}
-      step={1}
-      incrementButtonAriaLabel={'Increase value by 1'}
-      decrementButtonAriaLabel={'Decrease value by 1'}
-    />
-  </div>
-);
