--- packages/react-examples/src/react-next/SpinButton/docs/SpinButtonOverview.md
@@ -1,1 +1,0 @@
-A spin button (`SpinButton`) allows someone to incrementally adjust a value in small steps. It’s mainly used for numeric values, but other values are supported too.
