+++ packages/react-internal/src/components/SpinButton/SpinButton.test.tsx
@@ -1,15 +1,17 @@
 import * as React from 'react';
 import * as ReactTestUtils from 'react-dom/test-utils';
-import * as renderer from 'react-test-renderer';
+import { create } from '@uifabric/utilities/lib/test';
 import { ReactWrapper, mount } from 'enzyme';
-import { SpinButton, ISpinButtonState } from './SpinButton';
+
+import { SpinButton } from './SpinButton';
 import { ISpinButton, ISpinButtonProps } from './SpinButton.types';
 import { KeyCodes, resetIds } from '../../Utilities';
 import { mockEvent } from '../../common/testUtilities';
+import { isConformant } from '../../common/isConformant';
 
 describe('SpinButton', () => {
   let ref: React.RefObject<ISpinButton>;
-  let wrapper: ReactWrapper<ISpinButtonProps, ISpinButtonState, SpinButton> | undefined;
+  let wrapper: ReactWrapper<ISpinButtonProps> | undefined;
 
   beforeEach(() => {
     ref = React.createRef<ISpinButton>();
@@ -26,14 +28,19 @@ describe('SpinButton', () => {
     }
   });
 
+  isConformant({
+    Component: SpinButton,
+    displayName: 'SpinButton',
+  });
+
   it('renders correctly', () => {
-    const component = renderer.create(<SpinButton label="label" />);
+    const component = create(<SpinButton label="label" />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('renders correctly with user-provided values', () => {
-    const component = renderer.create(
+    const component = create(
       <SpinButton label="label" value="0" ariaValueNow={0} ariaValueText="0 pt" data-test="test" />,
     );
     const tree = component.toJSON();
@@ -56,8 +63,8 @@ describe('SpinButton', () => {
 
     const inputDOM = wrapper.getDOMNode().querySelector('input')!;
 
-    expect(inputDOM.getAttribute('aria-valuemin')).toBe(String(SpinButton.defaultProps.min));
-    expect(inputDOM.getAttribute('aria-valuemax')).toBe(String(SpinButton.defaultProps.max));
+    expect(inputDOM.getAttribute('aria-valuemin')).toBe('0');
+    expect(inputDOM.getAttribute('aria-valuemax')).toBe('100');
   });
 
   it('respects min and max in DOM', () => {
@@ -281,16 +288,32 @@ describe('SpinButton', () => {
     expect(ref.current!.value).toBe('1');
   });
 
-  it('accepts user-entered values', () => {
+  it('accepts user-entered values when uncontrolled', () => {
     wrapper = mount(<SpinButton componentRef={ref} defaultValue="12" />);
-
     const inputDOM = wrapper.getDOMNode().querySelector('input')!;
-    ReactTestUtils.Simulate.input(inputDOM, mockEvent('21'));
-    ReactTestUtils.Simulate.blur(inputDOM);
+
+    ReactTestUtils.act(() => {
+      ReactTestUtils.Simulate.input(inputDOM, mockEvent('21'));
+      ReactTestUtils.Simulate.blur(inputDOM);
+    });
 
     expect(ref.current!.value).toBe('21');
     expect(inputDOM.value).toBe('21');
     expect(inputDOM.getAttribute('aria-valuenow')).toBe('21');
+  });
+
+  it('does not accept user-entered values when controlled', () => {
+    wrapper = mount(<SpinButton componentRef={ref} value="12" />);
+
+    const inputDOM = wrapper.getDOMNode().querySelector('input')!;
+    ReactTestUtils.act(() => {
+      ReactTestUtils.Simulate.input(inputDOM, mockEvent('21'));
+      ReactTestUtils.Simulate.blur(inputDOM);
+    });
+
+    expect(ref.current!.value).toBe('12');
+    expect(inputDOM.value).toBe('12');
+    expect(inputDOM.getAttribute('aria-valuenow')).toBe('12');
   });
 
   it('resets value when user entry is invalid', () => {
@@ -403,19 +426,24 @@ describe('SpinButton', () => {
     wrapper = mount(<SpinButton defaultValue="12" />);
 
     const inputDOM = wrapper.getDOMNode().querySelector('input')!;
-    const buttonDOM = wrapper.getDOMNode().getElementsByClassName('ms-UpButton')[0];
+    const buttonDOM: any = wrapper.getDOMNode().querySelector('.ms-UpButton');
 
+    expect(inputDOM.value).toBe('12');
     expect(buttonDOM).toBeTruthy();
 
     // start spinning
-    ReactTestUtils.Simulate.mouseDown(buttonDOM, { type: 'mousedown', clientX: 0, clientY: 0 });
-    // spin again
-    jest.runOnlyPendingTimers();
-    // spin again
-    jest.runOnlyPendingTimers();
+    ReactTestUtils.act(() => {
+      ReactTestUtils.Simulate.mouseDown(buttonDOM, { type: 'mousedown', clientX: 0, clientY: 0 });
 
-    ReactTestUtils.Simulate.focus(inputDOM);
-    jest.runAllTimers();
+      // spin again
+      jest.runOnlyPendingTimers();
+
+      // spin again
+      jest.runOnlyPendingTimers();
+
+      ReactTestUtils.Simulate.focus(inputDOM);
+      jest.runAllTimers();
+    });
 
     const currentValue = inputDOM.value;
     expect(currentValue).not.toBe('12');
@@ -459,8 +487,8 @@ describe('SpinButton', () => {
     ReactTestUtils.Simulate.focus(inputDOM);
     ReactTestUtils.Simulate.input(inputDOM, mockEvent('99'));
     ReactTestUtils.Simulate.keyDown(inputDOM, { which: KeyCodes.enter });
-    expect(keyCode).toBe(KeyCodes.enter);
     expect(onValidate).toBeCalled();
+    expect(keyCode).toBe(KeyCodes.enter);
     ReactTestUtils.Simulate.blur(inputDOM);
     expect(onValidate).toHaveBeenCalledTimes(1);
   });
