--- packages/react-next/src/components/SpinButton/index.ts
@@ -1,2 +1,0 @@
-export * from './SpinButton';
-export * from './SpinButton.types';
